feat: Add ability to filter by time on timeline
